pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-takehome"
version = "0.1.0"
description = "Production-ready inference and data collection pipeline"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "Pillow>=10.0.0",
    "click>=8.1.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "tqdm>=4.66.0"
]

[project.scripts]
ml-takehome = "client.cli:cli"

[tool.black]
line-length = 88
target-version = ['py38']